pkg/sql/catalog: rename descriptortree to nstree
This ends up being a better name for where this is going.

Release note: None
