+++ pkg/BUILD.bazel
@@ -161,11 +161,11 @@ ALL_TESTS = [
     "//pkg/sql/catalog/colinfo:colinfo_test",
     "//pkg/sql/catalog/dbdesc:dbdesc_test",
     "//pkg/sql/catalog/descpb:descpb_test",
-    "//pkg/sql/catalog/descriptortree:descriptortree_test",
     "//pkg/sql/catalog/descs:descs_test",
     "//pkg/sql/catalog/hydratedtables:hydratedtables_test",
     "//pkg/sql/catalog/lease:lease_test",
     "//pkg/sql/catalog/multiregion:multiregion_test",
+    "//pkg/sql/catalog/nstree:nstree_test",
     "//pkg/sql/catalog/resolver:resolver_test",
     "//pkg/sql/catalog/schemadesc:schemadesc_test",
     "//pkg/sql/catalog/schemaexpr:schemaexpr_test",
