+++ pkg/sql/catalog/nstree/BUILD.bazel
@@ -1,13 +1,13 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
-    name = "descriptortree",
+    name = "nstree",
     srcs = [
         "by_id.go",
         "by_name.go",
         "tree.go",
     ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/sql/catalog/descriptortree",
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/catalog/nstree",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/sql/catalog",
@@ -18,10 +18,10 @@ go_library(
 )
 
 go_test(
-    name = "descriptortree_test",
+    name = "nstree_test",
     srcs = ["tree_test.go"],
     data = glob(["testdata/**"]),
-    embed = [":descriptortree"],
+    embed = [":nstree"],
     deps = [
         "//pkg/security",
         "//pkg/sql/catalog",
