+++ pkg/sql/catalog/nstree/by_name.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package descriptortree
+package nstree
 
 import (
 	"sync"
