+++ pkg/sql/catalog/nstree/testdata/desc_tree/iterate
